Separated shapefile test from goetiff test.
